+++ src/components/spartan/SCard/SCard.vue
@@ -1,8 +1,9 @@
 <script setup lang="ts">
 import { twMerge } from 'tailwind-merge';
 import { hasSlotContent, usePassthrough } from '@/helpers';
-import type { TCardProps } from './types';
 import { bodyStyles, containerStyles } from './styles';
+import { XMarkIcon } from '@heroicons/vue/24/outline';
+import type { TCardProps } from './types';
 import VIcon from './atoms/Icon.vue';
 import VActions from './atoms/Actions.vue';
 
@@ -11,20 +12,40 @@ const { pt, extractor } = usePassthrough();
 
 const [bodyClass, bodyProps] = extractor(pt.value.body);
 const [actionsClass, actionsProps] = extractor(pt.value.actions);
+const [titleClass, titleProps] = extractor(pt.value.title);
+const [descriptionClass, descriptionProps] = extractor(pt.value.description);
 const ptIcon = extractor(pt.value.icon);
 const ptIconContainer = extractor(pt.value.iconContainer);
 </script>
 
 <template>
     <article :class="twMerge(containerStyles({ size }), $props.class)">
-        <VIcon v-bind="{ icon, iconColor, iconType, ptIcon, ptIconContainer }" />
+        <header v-if="icon" :class="twMerge('mx-4 mb-8 mt-6 flex sm:mx-8 sm:mt-8', closable ? 'justify-between' : 'justify-center')">
+            <VIcon v-bind="{ icon, iconColor, iconType, ptIcon, ptIconContainer }" />
+            <XMarkIcon v-if="closable" class="h-5 w-5 text-gray-400" @click="$emit('close')" />
+        </header>
 
-        <main data-s-body v-bind="bodyProps" :class="twMerge(bodyStyles({ size }), bodyClass)">
-            <h3 v-if="hasSlotContent($slots.title)" class="text-center text-lg font-semibold text-gray-900">
-                <slot name="title" />
+        <main
+            data-s-body
+            v-bind="bodyProps"
+            :class="twMerge(bodyStyles({ size }), icon ? 'pt-0 sm:pt-0' : '', bodyClass)"
+        >
+            <h3
+                data-s-title
+                v-bind="titleProps"
+                v-if="hasSlotContent($slots.title) || title"
+                :class="twMerge('text-center text-base font-semibold text-gray-900', titleClass)"
+            >
+                <slot v-if="hasSlotContent($slots.title)" name="title" />
+                <template v-else>{{ title }}</template>
             </h3>
 
-            <p v-if="hasSlotContent($slots.description)" class="mt-2 text-center text-sm font-medium text-gray-500">
+            <p
+                data-s-description
+                v-bind="descriptionProps"
+                v-if="hasSlotContent($slots.description)"
+                :class="twMerge('mt-2 text-center text-sm font-normal text-gray-500', descriptionClass)"
+            >
                 <slot name="description" />
             </p>