+++ src/components/spartan/SModalConfirm/types.ts
@@ -1,0 +1,11 @@
+export type TModalConfirmEmits = {
+    (e: 'update:open', value: boolean): void;
+    (e: 'confirm'): void;
+    (e: 'cancel'): void;
+};
+
+export type TModalConfirmProps = {
+    text: string;
+    confirmText?: string;
+    cancelText?: string;
+};