+++ CHANGELOG.md
@@ -6,6 +6,16 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Added
+- `SModalConfirm` component.
+- `closable` prop to `SCard` and `SModalCard` components.
+
+### Changed
+- `SModalCard` component to use `v-model` for `open` prop.
+- New styles for `SCard`, `SModalCard` and `SModal` components.
+
+### Fixed
+- `SModal` responsive positioning.
 
 ## 2.2.1 - 2025-06-10
 ### Added